Extract UserService to Services Layer
- Create `services/user_service.py` module
- Update `api/auth.py` to import from services layer
- Update test imports
